feat: towards organization tests

+++ api/src/main/services/OrganizationService.ts
@@ -1,7 +1,7 @@
 import container from '../config/container';
 import { OrganizationApiKeyRole } from '../types/permissions';
 import OrganizationRepository from '../repositories/mongoose/OrganizationRepository';
-import { LeanApiKey, LeanOrganization, OrganizationFilter } from '../types/models/Organization';
+import { LeanApiKey, LeanOrganization, OrganizationFilter, OrganizationMember } from '../types/models/Organization';
 import { generateOrganizationApiKey } from '../utils/users/helpers';
 import UserService from './UserService';
 import { validateOrganizationData } from './validation/OrganizationServiceValidations';
@@ -50,13 +50,17 @@ class OrganizationService {
     };
   }
 
-  async create(organizationData: any): Promise<LeanOrganization> {
+  async create(organizationData: any, reqUser: any): Promise<LeanOrganization> {
     
     validateOrganizationData(organizationData);
     const proposedOwner = await this.userService.findByUsername(organizationData.owner);
 
     if (!proposedOwner) {
       throw new Error(`User with username ${organizationData.owner} does not exist.`);
+    }
+
+    if (proposedOwner.username !== reqUser.username && reqUser.role !== 'ADMIN') {
+      throw new Error('Only admins can create organizations for other users.');
     }
 
     organizationData = {
@@ -73,7 +77,18 @@ class OrganizationService {
     return organization;
   }
 
-  async addApiKey(organizationId: string, keyScope: OrganizationApiKeyRole): Promise<void> {
+  async addApiKey(organizationId: string, keyScope: OrganizationApiKeyRole, reqUser: any): Promise<void> {
+    
+    const organization = await this.organizationRepository.findById(organizationId);
+
+    if (!organization) {
+      throw new Error(`Organization with ID ${organizationId} does not exist.`);
+    }
+
+    if (organization.owner !== reqUser.username && reqUser.role !== 'ADMIN' && !organization.members.filter(m => m.username && ["OWNER", "ADMIN", "MANAGER"].includes(m.role)).map(m => m.username).includes(reqUser.username)) {
+      throw new Error('PERMISSION ERROR: Only SPACE admins or organization-level OWNER, ADMIN and MANAGER can add API keys to organizations they don\'t own.');
+    }
+    
     const apiKeyData: LeanApiKey = {
       key: generateOrganizationApiKey(),
       scope: keyScope
@@ -82,23 +97,41 @@ class OrganizationService {
     await this.organizationRepository.addApiKey(organizationId, apiKeyData);
   }
 
-  async addMember(organizationId: string, username: string): Promise<void> {
+  async addMember(organizationId: string, organizationMember: OrganizationMember, reqUser: any): Promise<void> {
     
-    const newMember = await this.userService.findByUsername(username);
+    if (!organizationMember.username || !organizationMember.role) {
+      throw new Error('INVALID DATA: organizationMember must have username and role.');
+    }
+
+    const organization = await this.organizationRepository.findById(organizationId);
+
+    if (!organization) {
+      throw new Error(`Organization with ID ${organizationId} does not exist.`);
+    }
+
+    if (organization.owner !== reqUser.username && reqUser.role !== 'ADMIN' && !organization.members.filter(m => m.username && ["OWNER", "ADMIN", "MANAGER"].includes(m.role)).map(m => m.username).includes(reqUser.username)) {
+      throw new Error('PERMISSION ERROR: Only SPACE admins or organization-level OWNER, ADMIN and MANAGER can add members to organizations they don\'t own.');
+    }
+
+    const newMember = await this.userService.findByUsername(organizationMember.username);
 
     if (!newMember) {
-      throw new Error(`User with username ${username} does not exist.`);
+      throw new Error(`User with username ${organizationMember.username} does not exist.`);
     }
     
-    await this.organizationRepository.addMember(organizationId, username);
+    await this.organizationRepository.addMember(organizationId, organizationMember);
   }
 
-  async update(organizationId: string, updateData: any): Promise<void> {
+  async update(organizationId: string, updateData: any, reqUser: any): Promise<void> {
     
     const organization = await this.organizationRepository.findById(organizationId);
 
     if (!organization) {
       throw new Error(`Organization with ID ${organizationId} does not exist.`);
+    }
+
+    if (organization.owner !== reqUser.username && reqUser.role !== 'ADMIN' && !organization.members.filter(m => m.username && ["OWNER", "ADMIN", "MANAGER"].includes(m.role)).map(m => m.username).includes(reqUser.username)) {
+      throw new Error('PERMISSION ERROR: Only SPACE admins or organization-level OWNER, ADMIN and MANAGER can update organizations.');
     }
 
     if (updateData.name) {
@@ -110,6 +143,10 @@ class OrganizationService {
     }
 
     if (updateData.owner) {
+      if (reqUser.role !== 'ADMIN' && organization.owner !== reqUser.username) {
+        throw new Error('PERMISSION ERROR: Only SPACE admins or organization owners can change organization ownership.');
+      }
+
       const proposedOwner = await this.userService.findByUsername(updateData.owner);
       if (!proposedOwner) {
         throw new Error(`INVALID DATA: User with username ${updateData.owner} does not exist.`);
@@ -121,11 +158,31 @@ class OrganizationService {
     await this.organizationRepository.update(organizationId, updateData);
   }
 
-  async removeApiKey(organizationId: string, apiKey: string): Promise<void> {
+  async removeApiKey(organizationId: string, apiKey: string, reqUser: any): Promise<void> {
+    const organization = await this.organizationRepository.findById(organizationId);
+    
+    if (!organization) {
+      throw new Error(`Organization with ID ${organizationId} does not exist.`);
+    }
+
+    if (organization.owner !== reqUser.username && reqUser.role !== 'ADMIN' && !organization.members.filter(m => m.username && ["OWNER", "ADMIN", "MANAGER"].includes(m.role)).includes(reqUser.username)) {
+      throw new Error('PERMISSION ERROR: Only SPACE admins or organization-level OWNER, ADMIN and MANAGER can remove API keys from organizations.');
+    }
+    
     await this.organizationRepository.removeApiKey(organizationId, apiKey);
   }
 
-  async removeMember(organizationId: string, username: string): Promise<void> {
+  async removeMember(organizationId: string, username: string, reqUser: any): Promise<void> {
+    const organization = await this.organizationRepository.findById(organizationId);
+
+    if (!organization) {
+      throw new Error(`Organization with ID ${organizationId} does not exist.`);
+    }
+
+    if (organization.owner !== reqUser.username && reqUser.role !== 'ADMIN' && !organization.members.filter(m => m.username && ["OWNER", "ADMIN", "MANAGER"].includes(m.role)).includes(reqUser.username)) {
+      throw new Error('PERMISSION ERROR: Only SPACE admins or organization-level OWNER, ADMIN and MANAGER can remove members from organizations.');
+    }
+    
     await this.organizationRepository.removeMember(organizationId, username);
   }
 }
